PHP 8.2 compatibility fix (#141)
* Body decode fix

In case of empty body (e.g. `HTTP 204` for DELETE). This `decodeBody`
regex fails and throws an error. The fix is to NOT decode an empty body.

* PHP 8.2 Compatibility fix

In Entity.php line 38:

During inheritance of ArrayAccess: Uncaught Return type of SaintSystems\OData\Entity::offsetGet($offset) should either be compatible with ArrayAccess::offsetGet(mixed $offset): mixed, or the #[\ReturnTypeWillChange] attribute should be used to temporarily suppress the notice

---------

Co-authored-by: Eugene Surov <eugene.surov@nic.at>